backend/requirements/base.txt
boto3==1.24.36
dj-database-url==0.5.0
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-filter==22.1
django-restframework==0.0.1
django-storages==1.12.3
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
pip-chill==1.0.1
psycopg2==2.9.3
whitenoise==6.2.0